extract artifactory implementations
